Fix progress calculation

+++ src/logdata/src/logfiltereddataworkerthread.cpp
@@ -31,6 +31,7 @@
 #include <moodycamel/blockingconcurrentqueue.h>
 
 #include <chrono>
+#include <cmath>
 
 namespace {
 struct PartialSearchResults {
@@ -407,7 +408,7 @@ void SearchOperation::doSearch( SearchData& searchData, LineNumber initialLine )
             }
 
             const int percentage 
-                = static_cast<int>( std::floor( 100.f * ( totalProcessedLines ).get() / totalLines.get() ) );
+                = static_cast<int>( std::floorf( 100.f * ( totalProcessedLines ).get() / totalLines.get() ) );
 
             if ( percentage > reportedPercentage || nbMatches > reportedMatches ) {
